Anh 18/07 5:46

+++ src/main/java/com/example/casestudy_movieproject/model/Vote.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDateTime;
+
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class Vote {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
-    private double score;
+    private int score;
     @ManyToOne
     @MapsId("userId")
     @JoinColumn(name ="user_Id")
